Cast Police
Eliminate some useless (but harmless) int<->size_t conversions.

+++ kernel/buffered.c
@@ -70,10 +70,10 @@ int X(toobig)(INT n)
 /* TRUE if there exists i < which such that maxnbuf[i] and
    maxnbuf[which] yield the same value, in which case we canonicalize
    on the minimum value */
-int X(nbuf_redundant)(INT n, INT vl, int which, 
-		      const INT *maxnbuf, int nmaxnbuf)
+int X(nbuf_redundant)(INT n, INT vl, size_t which, 
+		      const INT *maxnbuf, size_t nmaxnbuf)
 {
-     int i;
+     size_t i;
      (void)nmaxnbuf; /* UNUSED */
      for (i = 0; i < which; ++i)
 	  if (X(nbuf)(n, vl, maxnbuf[i]) == X(nbuf)(n, vl, maxnbuf[which]))
